+++ packages/generator/test/parser/v4/edmx-parser.spec.ts
@@ -21,8 +21,8 @@ describe('edmx-edmx-parser', () => {
     expect(parseEntityTypeV4(metadataEdmx.root).length).toBe(12);
     expect(parseFunctionImportsV4(metadataEdmx.root).length).toBe(8);
     expect(parseFunctions(metadataEdmx.root).length).toBe(8);
-    expect(parseActionImport(metadataEdmx.root).length).toBe(3);
-    expect(parseActions(metadataEdmx.root).length).toBe(3);
+    expect(parseActionImport(metadataEdmx.root).length).toBe(6);
+    expect(parseActions(metadataEdmx.root).length).toBe(5);
     expect(parseComplexTypesBase(metadataEdmx.root).length).toBe(4);
     expect(parseEnumTypes(metadataEdmx.root).length).toBe(2);