+++ test-packages/test-services/v4/test-service/action-imports.js
@@ -1,6 +1,6 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.actionImports = exports.testActionImportUnsupportedEdmTypes = exports.testActionImportMultipleParameterComplexReturnType = exports.testActionImportNoParameterNoReturnType = void 0;
+exports.actionImports = exports.testActionImportNoParameterEntityReturnType = exports.testActionImportUnsupportedEdmTypes = exports.testActionImportMultipleParameterComplexReturnType = exports.testActionImportNoParameterNoReturnType = void 0;
 /*
  * Copyright (c) 2020 SAP SE or an SAP affiliate company. All rights reserved.
  *
@@ -8,6 +8,7 @@ exports.actionImports = exports.testActionImportUnsupportedEdmTypes = exports.te
  */
 var core_1 = require("@sap-cloud-sdk/core");
 var TestComplexType_1 = require("./TestComplexType");
+var TestEntity_1 = require("./TestEntity");
 /**
  * Test Action Import No Parameter No Return Type.
  *
@@ -48,9 +49,21 @@ function testActionImportUnsupportedEdmTypes(parameters) {
     return new core_1.ActionImportRequestBuilder('/sap/opu/odata/sap/API_TEST_SRV', 'TestActionImportUnsupportedEdmTypes', function (data) { return core_1.transformReturnValueForEdmTypeV4(data, function (val) { return core_1.edmToTsV4(val.value, 'Edm.Any'); }); }, params);
 }
 exports.testActionImportUnsupportedEdmTypes = testActionImportUnsupportedEdmTypes;
+/**
+ * Test Action Import No Parameter Entity Return Type.
+ *
+ * @param parameters - Object containing all parameters for the action import.
+ * @returns A request builder that allows to overwrite some of the values and execute the resulting request.
+ */
+function testActionImportNoParameterEntityReturnType(parameters) {
+    var params = {};
+    return new core_1.ActionImportRequestBuilder('/sap/opu/odata/sap/API_TEST_SRV', 'TestActionImportNoParameterEntityReturnType', function (data) { return core_1.transformReturnValueForEntityV4(data, TestEntity_1.TestEntity); }, params);
+}
+exports.testActionImportNoParameterEntityReturnType = testActionImportNoParameterEntityReturnType;
 exports.actionImports = {
     testActionImportNoParameterNoReturnType: testActionImportNoParameterNoReturnType,
     testActionImportMultipleParameterComplexReturnType: testActionImportMultipleParameterComplexReturnType,
-    testActionImportUnsupportedEdmTypes: testActionImportUnsupportedEdmTypes
+    testActionImportUnsupportedEdmTypes: testActionImportUnsupportedEdmTypes,
+    testActionImportNoParameterEntityReturnType: testActionImportNoParameterEntityReturnType
 };
 //# sourceMappingURL=action-imports.js.map