+++ test-packages/test-services/v4/test-service/action-imports.ts
@@ -3,8 +3,9 @@
  *
  * This is a generated file powered by the SAP Cloud SDK for JavaScript.
  */
-import { transformReturnValueForUndefinedV4, transformReturnValueForComplexTypeV4, transformReturnValueForEdmTypeV4, edmToTsV4, deserializeComplexTypeV4, ActionImportRequestBuilder, ActionImportParameter } from '@sap-cloud-sdk/core';
+import { transformReturnValueForUndefinedV4, transformReturnValueForComplexTypeV4, transformReturnValueForEdmTypeV4, transformReturnValueForEntityV4, edmToTsV4, deserializeComplexTypeV4, ActionImportRequestBuilder, ActionImportParameter } from '@sap-cloud-sdk/core';
 import { TestComplexType } from './TestComplexType';
+import { TestEntity } from './TestEntity';
 
 /**
  * Type of the parameters to be passed to [[testActionImportNoParameterNoReturnType]].
@@ -89,8 +90,29 @@ export function testActionImportUnsupportedEdmTypes(parameters: TestActionImport
   return new ActionImportRequestBuilder('/sap/opu/odata/sap/API_TEST_SRV', 'TestActionImportUnsupportedEdmTypes', (data) => transformReturnValueForEdmTypeV4(data, (val) => edmToTsV4(val.value, 'Edm.Any')), params);
 }
 
+/**
+ * Type of the parameters to be passed to [[testActionImportNoParameterEntityReturnType]].
+ */
+export interface TestActionImportNoParameterEntityReturnTypeParameters {
+}
+
+/**
+ * Test Action Import No Parameter Entity Return Type.
+ *
+ * @param parameters - Object containing all parameters for the action import.
+ * @returns A request builder that allows to overwrite some of the values and execute the resulting request.
+ */
+export function testActionImportNoParameterEntityReturnType(parameters: TestActionImportNoParameterEntityReturnTypeParameters): ActionImportRequestBuilder<TestActionImportNoParameterEntityReturnTypeParameters, TestEntity> {
+  const params = {
+
+  }
+
+  return new ActionImportRequestBuilder('/sap/opu/odata/sap/API_TEST_SRV', 'TestActionImportNoParameterEntityReturnType', (data) => transformReturnValueForEntityV4(data, TestEntity), params);
+}
+
 export const actionImports = {
   testActionImportNoParameterNoReturnType,
   testActionImportMultipleParameterComplexReturnType,
-  testActionImportUnsupportedEdmTypes
+  testActionImportUnsupportedEdmTypes,
+  testActionImportNoParameterEntityReturnType
 };